fix some type issues

+++ Desktop_Interface/i2cdecoder.cpp
@@ -59,7 +59,7 @@ void i2cDecoder::run()
 
 int i2cDecoder::serialDistance(isoBuffer* buffer)
 {
-    int back_bit = buffer->m_back * 8;
+    uint64_t back_bit = buffer->m_back * 8;
     int bufferEnd_bit = buffer->m_bufferLen * 8;
     if (back_bit >= serialPtr_bit)
         return back_bit - serialPtr_bit;
